+++ packages/app-builder-lib/src/configuration.ts
@@ -21,7 +21,7 @@ import { NsisOptions, NsisWebOptions, PortableOptions } from "./targets/nsis/nsi
 /**
  * Configuration Options
  */
-export interface Configuration extends PlatformSpecificBuildOptions {
+export interface Configuration extends PlatformSpecificBuildOptions, Hooks {
   /**
    * The application id. Used as [CFBundleIdentifier](https://developer.apple.com/library/ios/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html#//apple_ref/doc/uid/20001431-102070) for MacOS and as
    * [Application User Model ID](https://msdn.microsoft.com/en-us/library/windows/desktop/dd378459(v=vs.85).aspx) for Windows (NSIS target only, Squirrel.Windows not supported). It is strongly recommended that an explicit ID is set.
@@ -162,7 +162,7 @@ export interface Configuration extends PlatformSpecificBuildOptions {
   /**
    * Returns the path to custom Electron build (e.g. `~/electron/out/R`). Zip files must follow the pattern `electron-v${version}-${platformName}-${arch}.zip`, otherwise it will be assumed to be an unpacked Electron app directory
    */
-  readonly electronDist?: string | ((options: PrepareApplicationStageDirectoryOptions) => string)
+  readonly electronDist?: string | CustomElectronDistributable
 
   /**
    * The [electron-download](https://github.com/electron-userland/electron-download#usage) options.
@@ -216,56 +216,6 @@ export interface Configuration extends PlatformSpecificBuildOptions {
   readonly framework?: string | null
 
   /**
-   * The function (or path to file or module id) to be [run before pack](#beforepack)
-   */
-  readonly beforePack?: ((context: BeforePackContext) => Promise<any> | any) | string | null
-
-  /**
-   * The function (or path to file or module id) to be [run after the prebuilt Electron binary has been extracted to the output directory](#afterextract)
-   */
-  readonly afterExtract?: ((context: AfterExtractContext) => Promise<any> | any) | string | null
-
-  /**
-   * The function (or path to file or module id) to be [run after pack](#afterpack) (but before pack into distributable format and sign).
-   */
-  readonly afterPack?: ((context: AfterPackContext) => Promise<any> | any) | string | null
-  /**
-   * The function (or path to file or module id) to be [run after pack and sign](#aftersign) (but before pack into distributable format).
-   */
-  readonly afterSign?: ((context: AfterPackContext) => Promise<any> | any) | string | null
-
-  /**
-   * The function (or path to file or module id) to be run on artifact build start.
-   */
-  readonly artifactBuildStarted?: ((context: ArtifactBuildStarted) => Promise<any> | any) | string | null
-  /**
-   * The function (or path to file or module id) to be run on artifact build completed.
-   */
-  readonly artifactBuildCompleted?: ((context: ArtifactCreated) => Promise<any> | any) | string | null
-  /**
-   * The function (or path to file or module id) to be [run after all artifacts are build](#afterAllArtifactBuild).
-   */
-  readonly afterAllArtifactBuild?: ((context: BuildResult) => Promise<Array<string>> | Array<string>) | string | null
-  /**
-   * MSI project created on disk - not packed into .msi package yet.
-   */
-  readonly msiProjectCreated?: ((path: string) => Promise<any> | any) | string | null
-  /**
-   * Appx manifest created on disk - not packed into .appx package yet.
-   */
-  readonly appxManifestCreated?: ((path: string) => Promise<any> | any) | string | null
-  /**
-   * The function (or path to file or module id) to be [run on each node module](#onnodemodulefile) file. Returning `true`/`false` will determine whether to force include or to use the default copier logic
-   */
-  readonly onNodeModuleFile?: ((path: string) => void | boolean) | string | null
-  /**
-   * The function (or path to file or module id) to be run before dependencies are installed or rebuilt. Works when `npmRebuild` is set to `true`. Resolving to `false` will skip dependencies install or rebuild.
-   *
-   * If provided and `node_modules` are missing, it will not invoke production dependencies check.
-   */
-  readonly beforeBuild?: ((context: BeforeBuildContext) => Promise<any>) | string | null
-
-  /**
    * Whether to include PDB files.
    * @default false
    */
@@ -292,7 +242,11 @@ export interface Configuration extends PlatformSpecificBuildOptions {
   readonly disableSanityCheckAsar?: boolean
 }
 
-interface PackContext {
+export type CustomElectronDistributable = (options: PrepareApplicationStageDirectoryOptions) => string
+
+export type Hook<T, V> = (contextOrPath: T) => Promise<V> | V
+
+export interface PackContext {
   readonly outDir: string
   readonly appOutDir: string
   readonly packager: PlatformPackager<any>
@@ -304,6 +258,102 @@ export type AfterPackContext = PackContext
 export type BeforePackContext = PackContext
 export type AfterExtractContext = PackContext
 
+export interface Hooks {
+  /**
+The function (or path to file or module id) to be run before pack.
+
+```typescript
+(context: BeforePackContext): Promise<any> | any
+```
+
+!!! example "As function"
+
+    ```js
+    beforePack: async (context) => {
+      // your code
+    }
+    ```
+
+Because in a configuration file you cannot use JavaScript, can be specified as a path to file or module id. Function must be exported as default export.
+
+```json
+"build": {
+  "beforePack": "./myBeforePackHook.js"
+}
+```
+
+File `myBeforePackHook.js` in the project root directory:
+
+!!! example "myBeforePackHook.js"
+    ```js
+    exports.default = async function(context) {
+      // your custom code
+    }
+    ```
+   */
+  readonly beforePack?: Hook<BeforePackContext, any> | string | null
+
+  /**
+   * The function (or path to file or module id) to be [run after the prebuilt Electron binary has been extracted to the output directory](#afterextract)
+   */
+  readonly afterExtract?: Hook<AfterExtractContext, any> | string | null
+
+  /**
+   * The function (or path to file or module id) to be [run after pack](#afterpack) (but before pack into distributable format and sign).
+   */
+  readonly afterPack?: Hook<AfterPackContext, any> | string | null
+
+  /**
+   * The function (or path to file or module id) to be [run after pack and sign](#aftersign) (but before pack into distributable format).
+   */
+  readonly afterSign?: Hook<AfterPackContext, any> | string | null
+
+  /**
+   * The function (or path to file or module id) to be run on artifact build start.
+   */
+  readonly artifactBuildStarted?: Hook<ArtifactBuildStarted, any> | string | null
+  /**
+   * The function (or path to file or module id) to be run on artifact build completed.
+   */
+  readonly artifactBuildCompleted?: Hook<ArtifactCreated, any> | string | null
+  /**
+   * The function (or path to file or module id) to be run after all artifacts are built.
+
+```typescript
+(buildResult: BuildResult): Promise<Array<string>> | Array<string>
+```
+
+Configuration in the same way as `afterPack` (see above).
+
+!!! example "myAfterAllArtifactBuild.js"
+    ```js
+    exports.default = function () {
+      // you can return additional files to publish
+      return ["/path/to/additional/result/file"]
+    }
+    ```
+   */
+  readonly afterAllArtifactBuild?: Hook<BuildResult, Array<string>> | string | null
+  /**
+   * MSI project created on disk - not packed into .msi package yet.
+   */
+  readonly msiProjectCreated?: Hook<string, any> | string | null
+  /**
+   * Appx manifest created on disk - not packed into .appx package yet.
+   */
+  readonly appxManifestCreated?: Hook<string, any> | string | null
+  /**
+   * The function (or path to file or module id) to be [run on each node module](#onnodemodulefile) file. Returning `true`/`false` will determine whether to force include or to use the default copier logic
+   */
+  readonly onNodeModuleFile?: Hook<string, void | boolean> | string | null
+  /**
+   * The function (or path to file or module id) to be run before dependencies are installed or rebuilt. Works when `npmRebuild` is set to `true`. Resolving to `false` will skip dependencies install or rebuild.
+   *
+   * If provided and `node_modules` are missing, it will not invoke production dependencies check.
+   */
+  readonly beforeBuild?: Hook<BeforeBuildContext, boolean | void> | string | null
+}
+
 export interface MetadataDirectories {
   /**
    * The path to build resources.
@@ -314,7 +364,7 @@ export interface MetadataDirectories {
   readonly buildResources?: string | null
 
   /**
-   * The output directory. [File macros](/file-patterns#file-macros) are supported.
+   * The output directory. [File macros](./file-patterns.md#file-macros) are supported.
    * @default dist
    */
   readonly output?: string | null