+++ packages/app-builder-lib/src/publish/s3/spacesPublisher.ts
@@ -16,10 +16,10 @@ export default class SpacesPublisher extends BaseS3Publisher {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   static checkAndResolveOptions(options: SpacesOptions, channelFromAppVersion: string | null, errorIfCannot: boolean) {
     if (options.name == null) {
-      throw new InvalidConfigurationError(`Please specify "name" for "spaces" publish provider (see https://www.electron.build/configuration/publish#spacesoptions)`)
+      throw new InvalidConfigurationError(`Please specify "name" for "spaces" publish provider (see https://www.electron.build/publish#spacesoptions)`)
     }
     if (options.region == null) {
-      throw new InvalidConfigurationError(`Please specify "region" for "spaces" publish provider (see https://www.electron.build/configuration/publish#spacesoptions)`)
+      throw new InvalidConfigurationError(`Please specify "region" for "spaces" publish provider (see https://www.electron.build/publish#spacesoptions)`)
     }
 
     if (options.channel == null && channelFromAppVersion != null) {
@@ -41,10 +41,10 @@ export default class SpacesPublisher extends BaseS3Publisher {
     const accessKey = process.env.DO_KEY_ID
     const secretKey = process.env.DO_SECRET_KEY
     if (isEmptyOrSpaces(accessKey)) {
-      throw new InvalidConfigurationError("Please set env DO_KEY_ID (see https://www.electron.build/configuration/publish#spacesoptions)")
+      throw new InvalidConfigurationError("Please set env DO_KEY_ID (see https://www.electron.build/publish#spacesoptions)")
     }
     if (isEmptyOrSpaces(secretKey)) {
-      throw new InvalidConfigurationError("Please set env DO_SECRET_KEY (see https://www.electron.build/configuration/publish#spacesoptions)")
+      throw new InvalidConfigurationError("Please set env DO_SECRET_KEY (see https://www.electron.build/publish#spacesoptions)")
     }
     args.push("--accessKey", accessKey)
     args.push("--secretKey", secretKey)