+++ packages/app-builder-lib/src/util/appFileCopier.ts
@@ -197,24 +197,25 @@ export async function computeNodeModuleFileSets(platformPackager: PlatformPackag
     // use main matcher patterns, so, user can exclude some files !node_modules/xxxx
     return path.dirname(parentDir)
   }
-  for (const info of deps) {
-    const source = info.dir
-    const destination = path.join(mainMatcher.to, NODE_MODULES, info.name)
+  const collectNodeModules = async (dep: NodeModuleInfo, destination: string) => {
+    const source = dep.dir
     const matcher = new FileMatcher(getRealSource(source), destination, mainMatcher.macroExpander, mainMatcher.patterns)
     const copier = new NodeModuleCopyHelper(matcher, platformPackager.info)
-    const files = await copier.collectNodeModules(info, nodeModuleExcludedExts)
+    const files = await copier.collectNodeModules(dep, nodeModuleExcludedExts)
     result[index++] = validateFileSet({ src: source, destination, files, metadata: copier.metadata })
 
-    if (info.conflictDependency) {
-      for (const dep of info.conflictDependency) {
-        const source = dep.dir
-        const destination = path.join(mainMatcher.to, NODE_MODULES, info.name, NODE_MODULES, dep.name)
-        const matcher = new FileMatcher(getRealSource(source), destination, mainMatcher.macroExpander, mainMatcher.patterns)
-        const copier = new NodeModuleCopyHelper(matcher, platformPackager.info)
-        result[index++] = validateFileSet({ src: source, destination, files: await copier.collectNodeModules(dep, nodeModuleExcludedExts), metadata: copier.metadata })
+    if (dep.conflictDependency) {
+      for (const c of dep.conflictDependency) {
+        await collectNodeModules(c, path.join(destination, NODE_MODULES, c.name))
       }
     }
   }
+
+  for (const dep of deps) {
+    const destination = path.join(mainMatcher.to, NODE_MODULES, dep.name)
+    await collectNodeModules(dep, destination)
+  }
+
   return result
 }