+++ packages/electron-updater/src/AppUpdater.ts
@@ -10,6 +10,7 @@ import {
   CancellationError,
   ProgressInfo,
   BlockMap,
+  retry,
 } from "builder-util-runtime"
 import { randomBytes } from "crypto"
 import { release } from "os"
@@ -256,7 +257,7 @@ export abstract class AppUpdater extends (EventEmitter as new () => TypedEmitter
   }
 
   /**
-   * Configure update provider. If value is `string`, [GenericServerOptions](/configuration/publish#genericserveroptions) will be set with value as `url`.
+   * Configure update provider. If value is `string`, [GenericServerOptions](./publish.md#genericserveroptions) will be set with value as `url`.
    * @param options If you want to override configuration in the `app-update.yml`.
    */
   setFeedURL(options: PublishConfiguration | AllPublishOptions | string) {
@@ -712,7 +713,14 @@ export abstract class AppUpdater extends (EventEmitter as new () => TypedEmitter
     const tempUpdateFile = await createTempUpdateFile(`temp-${updateFileName}`, cacheDir, log)
     try {
       await taskOptions.task(tempUpdateFile, downloadOptions, packageFile, removeFileIfAny)
-      await rename(tempUpdateFile, updateFile)
+      await retry(
+        () => rename(tempUpdateFile, updateFile),
+        60,
+        500,
+        0,
+        0,
+        error => error instanceof Error && /^EBUSY:/.test(error.message)
+      )
     } catch (e: any) {
       await removeFileIfAny()