+++ test/src/BuildTest.ts
@@ -1,5 +1,5 @@
 import { checkBuildRequestOptions } from "app-builder-lib"
-import { doMergeConfigs } from "app-builder-lib/out/util/config"
+import { doMergeConfigs } from "app-builder-lib/out/util/config/config"
 import { Arch, createTargets, DIR_TARGET, Platform } from "electron-builder"
 import { promises as fs } from "fs"
 import { outputJson } from "fs-extra"